requirements.txt.txt
aiofiles==24.1.0
aiogram==3.16.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.3.0
captcha==0.6.0
certifi==2024.8.30
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
pillow==11.0.0
propcache==0.2.1
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==21.7
rsa==4.9
sniffio==1.3.1
Telethon==1.36.0
typing_extensions==4.12.2
yarl==1.18.3